requirements.txt
numpy==1.26.4
pymavlink
Flask==3.0.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
eventlet
google-generativeai==0.4.1
pyserial==3.5
pyaudio==0.2.14
google-cloud-speech==2.26.0
Flask-Cors==4.0.1